Randos/multi-agent-orchestration_legacy/main.py
#!/usr/bin/env python3
"""
Multi-Agent AI Orchestration System
Integrates CrewAI, LangChain, Mem0, and environmental monitoring
"""

import os
import asyncio
from datetime import datetime
from typing import List, Dict, Any

# Core orchestration imports
from crewai import Agent, Task, Crew, Process
from langchain_openai import ChatOpenAI
from langchain.memory import ConversationBufferMemory
from mem0 import Memory

# Workflow orchestration
from prefect import flow, task

# Experiment tracking
import mlflow

# Environmental monitoring
from environmental.environmental_monitor import EnvironmentalMonitor, BiodiversityTracker

# Configuration
os.environ["OPENAI_API_KEY"] = "your-api-key-here"  # Replace with actual key

class MultiAgentOrchestrator:
    """Main orchestration system for multi-agent AI with environmental focus"""
    
    def __init__(self):
        # Initialize memory system
        self.memory = Memory()
        
        # Initialize LLM
        self.llm = ChatOpenAI(
            model="gpt-4",
            temperature=0.7
        )
        
        # Initialize conversation memory
        self.conversation_memory = ConversationBufferMemory()
        
        # Initialize environmental systems
        self.env_monitor = EnvironmentalMonitor()
        self.bio_tracker = BiodiversityTracker()
        
        # Initialize agents
        self.initialize_agents()
    
    def initialize_agents(self):
        """Initialize specialized CrewAI agents"""
        
        # Environmental Analysis Agent
        self.env_analyst = Agent(
            role='Environmental Data Analyst',
            goal='Analyze environmental sensor data and identify patterns',
            backstory="""You are an expert in environmental science with 
            specialization in real-time sensor data analysis and pattern recognition.""",
            verbose=True,
            allow_delegation=False,
            llm=self.llm
        )
        
        # Biodiversity Conservation Agent
        self.bio_conservationist = Agent(
            role='Biodiversity Conservation Specialist',
            goal='Monitor species populations and recommend conservation actions',
            backstory="""You are a conservation biologist focused on 
            protecting marine ecosystems and endangered species.""",
            verbose=True,
            allow_delegation=False,
            llm=self.llm
        )
        
        # Research Coordinator Agent
        self.research_coordinator = Agent(
            role='Research Coordinator',
            goal='Coordinate research efforts and synthesize findings',
            backstory="""You coordinate multi-disciplinary research teams 
            and ensure effective collaboration between different specialists.""",
            verbose=True,
            allow_delegation=True,
            llm=self.llm
        )
        
        # Emergency Response Agent
        self.emergency_responder = Agent(
            role='Environmental Emergency Responder',
            goal='Detect and respond to environmental emergencies',
            backstory="""You are trained in rapid response to environmental 
            crises including pollution events and habitat destruction.""",
            verbose=True,
            allow_delegation=False,
            llm=self.llm
        )
    
    @task
    def collect_environmental_data(self) -> List[Dict]:
        """Prefect task to collect environmental data"""
        print("Collecting environmental data...")
        # Simulate data collection
        data = []
        for sensor_type in ["temperature", "co2", "humidity", "water_quality"]:
            data.append({
                "type": sensor_type,
                "value": np.random.random() * 100,
                "timestamp": datetime.now().isoformat()
            })
        return data
    
    @task
    def analyze_biodiversity(self) -> Dict:
        """Prefect task to analyze biodiversity metrics"""
        print("Analyzing biodiversity metrics...")
        return {
            "diversity_index": 2.45,
            "species_count": 15,
            "habitat_health": "Good",
            "timestamp": datetime.now().isoformat()
        }
    
    @flow(name="Environmental Monitoring Flow")
    def environmental_monitoring_flow(self):
        """Main Prefect flow for environmental monitoring"""
        # Collect data
        env_data = self.collect_environmental_data()
        
        # Analyze biodiversity
        bio_metrics = self.analyze_biodiversity()
        
        # Log to MLflow
        with mlflow.start_run():
            mlflow.log_metrics({
                "diversity_index": bio_metrics["diversity_index"],
                "species_count": bio_metrics["species_count"]
            })
            
            for data_point in env_data:
                mlflow.log_metric(f"{data_point['type']}_value", data_point['value'])
        
        return {
            "environmental_data": env_data,
            "biodiversity_metrics": bio_metrics
        }
    
    async def run_multi_agent_analysis(self, data: Dict) -> str:
        """Run multi-agent analysis using CrewAI"""
        
        # Define tasks for the crew
        data_analysis_task = Task(
            description=f"""Analyze the following environmental data and identify 
            any concerning patterns or anomalies: {data['environmental_data']}""",
            agent=self.env_analyst,
            expected_output="Environmental data analysis report with identified patterns"
        )
        
        biodiversity_task = Task(
            description=f"""Review the biodiversity metrics and recommend 
            conservation actions: {data['biodiversity_metrics']}""",
            agent=self.bio_conservationist,
            expected_output="Conservation recommendations based on biodiversity analysis"
        )
        
        synthesis_task = Task(
            description="""Synthesize the environmental analysis and conservation 
            recommendations into a comprehensive action plan.""",
            agent=self.research_coordinator,
            expected_output="Comprehensive environmental action plan"
        )
        
        # Create and run the crew
        crew = Crew(
            agents=[self.env_analyst, self.bio_conservationist, self.research_coordinator],
            tasks=[data_analysis_task, biodiversity_task, synthesis_task],
            process=Process.sequential,
            verbose=True
        )
        
        result = crew.kickoff()
        
        # Store results in memory
        self.memory.add(
            messages=[
                {
                    "role": "system",
                    "content": f"Environmental analysis completed at {datetime.now()}"
                },
                {
                    "role": "assistant",
                    "content": str(result)
                }
            ],
            user_id="environmental_system"
        )
        
        return str(result)
    
    async def detect_emergencies(self, data: Dict) -> List[str]:
        """Detect potential environmental emergencies"""
        emergencies = []
        
        for data_point in data.get('environmental_data', []):
            # Simple threshold-based detection
            if data_point['type'] == 'co2' and data_point['value'] > 80:
                emergencies.append(f"HIGH CO2 ALERT: {data_point['value']:.2f}")
            elif data_point['type'] == 'water_quality' and data_point['value'] < 20:
                emergencies.append(f"POOR WATER QUALITY: {data_point['value']:.2f}")
        
        if emergencies:
            # Create emergency response task
            emergency_task = Task(
                description=f"""Respond to the following environmental emergencies: 
                {', '.join(emergencies)}. Provide immediate action recommendations.""",
                agent=self.emergency_responder,
                expected_output="Emergency response plan with immediate actions"
            )
            
            emergency_crew = Crew(
                agents=[self.emergency_responder],
                tasks=[emergency_task],
                process=Process.sequential
            )
            
            response = emergency_crew.kickoff()
            print(f"⚠️  EMERGENCY RESPONSE: {response}")
        
        return emergencies
    
    async def run_orchestration_cycle(self):
        """Run a complete orchestration cycle"""
        print("="*50)
        print("Starting Multi-Agent Orchestration Cycle")
        print("="*50)
        
        # Step 1: Collect environmental data
        print("\n1. Running environmental monitoring flow...")
        data = self.environmental_monitoring_flow()
        
        # Step 2: Check for emergencies
        print("\n2. Checking for environmental emergencies...")
        emergencies = await self.detect_emergencies(data)
        
        if emergencies:
            print(f"   Found {len(emergencies)} emergency conditions!")
        else:
            print("   No emergencies detected.")
        
        # Step 3: Run multi-agent analysis
        print("\n3. Running multi-agent analysis...")
        analysis_result = await self.run_multi_agent_analysis(data)
        
        # Step 4: Retrieve relevant memories
        print("\n4. Retrieving relevant memories...")
        memories = self.memory.search(
            query="environmental analysis",
            user_id="environmental_system",
            limit=5
        )
        
        if memories and 'results' in memories:
            print(f"   Found {len(memories['results'])} relevant memories")
        
        # Step 5: Generate summary
        print("\n5. Generating orchestration summary...")
        summary = {
            "timestamp": datetime.now().isoformat(),
            "data_collected": len(data.get('environmental_data', [])),
            "emergencies": len(emergencies),
            "biodiversity_status": data['biodiversity_metrics']['habitat_health'],
            "analysis_completed": True
        }
        
        print("\n" + "="*50)
        print("Orchestration Cycle Complete")
        print(f"Summary: {summary}")
        print("="*50)
        
        return summary

async def main():
    """Main entry point"""
    # Initialize MLflow tracking
    mlflow.set_experiment("multi-agent-environmental-monitoring")
    
    # Create orchestrator
    orchestrator = MultiAgentOrchestrator()
    
    # Run orchestration cycle
    try:
        result = await orchestrator.run_orchestration_cycle()
        print(f"\nFinal Result: {result}")
    except Exception as e:
        print(f"Error during orchestration: {e}")
        print("Note: Please set your OPENAI_API_KEY environment variable to use AI features")

if __name__ == "__main__":
    # Note: numpy import for simulation
    import numpy as np
    
    print("""
    ╔══════════════════════════════════════════════════════╗
    ║   Multi-Agent AI Orchestration System               ║
    ║   Environmental Monitoring & Conservation Platform   ║
    ╚══════════════════════════════════════════════════════╝
    """)
    
    asyncio.run(main())